requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic>=2.6
anthropic>=0.40
google-genai>=0.3
openai>=1.50
pypdf>=4.3
json-repair>=0.25
